Add OAM Update Delay video test
The hardware is always rendering sprites to the linebuffer 1 line ahead
of the current scanline. Therefore, there is a 1-scanline delay between
when an OAM update is made and when it's visible. This test demonstrates
that by jagging a single 64x64 sprite by changing its position during
HBlank at the same time as the BG color.